feat(#115): move BytecodeClass, BytecodeMethod and BytecodeInstruction into separate package
